Version 0.0.5-hotfix
- HOTFIX
The blockchain had been corrupted when importing a few blocks of a node from the testnet in the mainnet.
Now it is controlled in all the sites that the main block is the same as the peer where we synchronize.
In addition, it is explicitly controlled that each block to be imported points to the previous one (Before we trusted a little in the peer information)

- Added checking balance on receive transaction
Now each node, before propagating the pending transaction to the network, will check if said wallet has the necessary balance to make that transaction. In this way we prevent a node from modifying the balance information of an account and can issue transactions to the network

- Improved Hashrate feedback
Now regardless of the speed of the miner should get a more real statistics

- Added blacklist of peers
If when we try to synchronize information from a peer we see that the information is not secure, it is failing to validate the blocks, the block does not match the previous one etc. we will block the peer for 10 min to prevent our local blockchain from being corrupted