Version 0.0.7
- API JSON-RPC / HTTP
With this api, you can create, visualize and manage the wallets and information of the blockchain
If a node is off, it will not respond to requests made to the api or the gossip

- Fixed propagation transactions
The system of prevention of modified balances was blocking the propagation of pending transactions.
Now they are sent correctly and prevent transactions that do not have balance in the account from being sent

- Discord link
If WEBHOOK_DISCORD is set on CONFIG.PHP, the node will send a message to discord using the WEBHOOK defined

- Some code fixes
The miner parameter in the documentation has been corrected.
The VERSION variable of CONFIG.php has been moved to another document for easy updating
Corrected some messages of log for a better identification
Added some parameters for the difficulty to work correctly in the testnet. The testnet network has an average of 2min per block