Forgotten transactions
BEWARE: Do not commit as is, would be a breaking change, this needs hard fork.

Spotted a typo (2 places). As a result, only the last transaction hash was embedded in the message, not all block transactions.
Consequences to be asserted with care.

A compatibility patch can be done by first changing isValid() to test for both messages (all transactions or just the last one) when there is more than the coinbase transaction in the block.
Then, after a while, changing the mine() function.
This could allow to do it in a smooth way.